requirements.txt
tqdm
nltk
rust-fst
scikit-learn
scipy
faiss-cpu
numpy